reverted to Long key in week dates

+++ src/main/java/com/softserve/edu/manager/ScheduleRowsManager.java
@@ -28,7 +28,7 @@ public interface ScheduleRowsManager {
      *
      * @return Map<Long ,   String>.
      */
-    Map<Byte, String> getWeekHead();
+    Map<Long, String> getWeekHead();
 
     void setCalendar(Calendar calendar);
 }
